Update: Post DELETE API protection

+++ app/api/posts/[postId]/route.ts
@@ -1,5 +1,7 @@
 import { NextResponse } from "next/server";
 import prisma from "@/prisma/prisma";
+import { authOptions } from "@/app/api/auth/[...nextauth]/route";
+import { getServerSession } from "next-auth/next";
 
 //here
 
@@ -81,13 +83,48 @@ export async function PUT(
   }
 }
 
-//here
-
 export async function DELETE(
   request: Request,
   { params }: { params: { postId: string } }
 ) {
+  const session = await getServerSession(authOptions);
+
   try {
+    if (!session) {
+      return NextResponse.json(
+        { error: "You are not logged in" },
+        { status: 401 }
+      );
+    }
+
+    const userId = session.user.id;
+    const user = await prisma.user.findUnique({
+      where: {
+        id: userId,
+      },
+    });
+
+    if (!user) {
+      return NextResponse.json({ error: "User not found" }, { status: 404 });
+    }
+
+    const comment = await prisma.comment.findUnique({
+      where: {
+        id: params.postId,
+      },
+    });
+
+    if (!comment) {
+      return NextResponse.json({ error: "Post not found" }, { status: 404 });
+    }
+
+    if (comment.userId !== userId) {
+      return NextResponse.json(
+        { error: "You are not authorized to delete this post" },
+        { status: 403 }
+      );
+    }
+
     const post = await prisma.post.delete({
       where: {
         id: params.postId,
